Shortened slurm_resume
Commented out unnecessary ip-mangling bits at the end of slurm_resume,
in addition to extra ansible tasks. Greatly improves the speed of
the script, and now finishes *before* a job is actually started
on the node. Previously all this stuff would run after the node
was actually responding to slurmctld, causing all sorts of issues.

Signed-off-by: Eric Coulter <jecoulte@iu.edu>